LayoutTests/inspector/formatting/formatting-xml-expected.txt

Test XML formatting.


== Running test suite: HTMLFormatter.XML
-- Running test case: HTMLFormatter.XML
PASS: atom.xml
PASS: basic.xml
PASS: rss.xml
PASS: valid-html-invalid-xml.xml
PASS: xslt.xml
